# shared/contracts/openapi.nexus-core.snapshot.yaml
openapi: 3.0.3
info:
  title: Nexus Core API
  version: "1.1.0"
  description: |
    Multi-tenant Agent Tool Gateway for controlled tool execution.

    Authentication supports API key (`X-NEXUS-GATEWAY-KEY`) and optional JWT bearer mode (`Authorization: Bearer`) when enabled by feature flags.

    Idempotency semantics for `POST /v1/run` on WRITE tools:
    - Same `Idempotency-Key` + same request fingerprint + `COMPLETED` => replay cached success.
    - Same `Idempotency-Key` + same fingerprint + `IN_PROGRESS` => `IDEMPOTENCY_IN_PROGRESS`.
    - Same `Idempotency-Key` + different fingerprint => `IDEMPOTENCY_KEY_CONFLICT`.
    - Same `Idempotency-Key` + same fingerprint + `FAILED` => terminal replay of cached error.
    - To retry after `FAILED`, clients MUST use a new `Idempotency-Key`.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/tools:
    get:
      tags: [Tools]
      operationId: listTools
      summary: List tenant tools
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XNexusActor"
        - $ref: "#/components/parameters/XNexusRole"
        - $ref: "#/components/parameters/XNexusScopes"
      responses:
        "200":
          description: Tools list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Tools]
      operationId: createTool
      summary: Create tool
      security:
        - NexusApiKey: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolCreateRequest"
      responses:
        "201":
          description: Tool created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "400":
          description: Validation or schema error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/tools/{name}:
    get:
      tags: [Tools]
      operationId: getToolByName
      summary: Get tool by name
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      responses:
        "200":
          description: Tool detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Tool not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Tools]
      operationId: updateToolByName
      summary: Partial update tool (name immutable)
      security:
        - NexusApiKey: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ToolNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolUpdateRequest"
      responses:
        "200":
          description: Tool updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    NexusApiKey:
      type: apiKey
      in: header
      name: X-NEXUS-GATEWAY-KEY
      description: Tenant API key (stored hashed in DB).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional JWT+JWKS auth mode when enabled by feature flags.
  parameters:
    ToolNamePath:
      name: name
      in: path
      required: true
      schema:
        type: string
    XNexusActor:
      name: X-NEXUS-ACTOR
      in: header
      required: false
      schema:
        type: string
      description: Delegated actor identifier for audit/policy context.
    XNexusRole:
      name: X-NEXUS-ROLE
      in: header
      required: false
      schema:
        type: string
      description: Delegated role for policy context.
    XNexusScopes:
      name: X-NEXUS-SCOPES
      in: header
      required: false
      schema:
        type: string
      description: Comma-separated scopes; intersected with key/JWT scopes.
...
  schemas:
    ErrorObject:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        error:
          $ref: "#/components/schemas/ErrorObject"
      required: [request_id, error]
...
    Tool:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        kind:
          type: string
          enum: [http]
        description:
          type: string
          nullable: true
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        url:
          type: string
        input_schema:
          type: object
          additionalProperties: true
        output_schema:
          type: object
          additionalProperties: true
          nullable: true
        action_type:
          type: string
          enum: [read, write]
        classification:
          type: string
          enum: [internal, external]
        sensitivity:
          type: string
          enum: [low, medium, high]
        risk_level:
          type: integer
          minimum: 1
          maximum: 5
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - kind
        - method
        - url
        - input_schema
        - action_type
        - classification
        - sensitivity
        - risk_level
        - enabled
        - created_at
        - updated_at
    ToolCreateRequest:
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [http]
        description:
          type: string
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        url:
          type: string
        input_schema:
          type: object
          additionalProperties: true
        output_schema:
          type: object
          additionalProperties: true
        action_type:
          type: string
          enum: [read, write]
        classification:
          type: string
          enum: [internal, external]
          default: internal
        sensitivity:
          type: string
          enum: [low, medium, high]
          default: low
        risk_level:
          type: integer
          minimum: 1
          maximum: 5
        enabled:
          type: boolean
      required: [name, kind, method, url, input_schema, action_type, risk_level, enabled]
    ToolUpdateRequest:
      type: object
      properties:
        description:
          type: string
          nullable: true
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        url:
          type: string
        input_schema:
          type: object
          additionalProperties: true
        output_schema:
          type: object
          additionalProperties: true
          nullable: true
        action_type:
          type: string
          enum: [read, write]
        classification:
          type: string
          enum: [internal, external]
        sensitivity:
          type: string
          enum: [low, medium, high]
        risk_level:
          type: integer
          minimum: 1
          maximum: 5
        enabled:
          type: boolean
    ToolListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
      required: [items]
...